One of my assignments in CSE 340
Intended idea for the assignment was to be able to parse a given program written in the grammar of the language that my professor had made up. We parse the program and were able to resolve the individual variables to its according scope, depending on if it was private or public.